requirements.txt
django==2.1.5
django-bootstrap4
django-allauth==0.38.0
django-extensions
Werkzeug
pyOpenSSL
